minor: resolve #537 conversations
